+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [1.3.10] - 12/09/2016
 - Refactor Cloudfront distributions
+- Added an Output for the endpoint of each deployment that is created for api gateway units. Storing these in an 'endpoints' list.
 
 ## [1.3.9] - 8/09/2016
 - Add extra configuration for ELBs - healthy/unhealthy thresholds, interval, timeout